Make logrotate::file more resilient
Validate that the first option is a string.
Mutate the path to remove /'s, allowing full paths to be passed.

Prior to this, it looks like we are passing basenames to logrotate and
the logrotate daemon is just nice enough to add /var/log/ to the
beginning of the file to be rotated. I don't see this documented
anywhere in the logrotate docs.

Change-Id: I341290c430023e5e9a24bf2247482ed818ea4a26